django_backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backports.zoneinfo==0.2.1
boto3==1.23.2
botocore==1.26.2
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
click==8.1.3
configparser==5.2.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
msgpack==1.0.4
mysql-connector-python==8.0.29
mysqlclient==2.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
zope.interface==5.4.0
